+++ source/Public/Get-GitLocalBranchName.ps1
@@ -4,26 +4,49 @@
 
     .DESCRIPTION
         The Get-GitLocalBranchName command is used to retrieve the name of the
-        current local Git branch. It uses the `git rev-parse --abbrev-ref HEAD`
-        command to get the branch name.
+        local Git branch. It can either return the name of the current branch or
+        search for branches based on a specified name or wildcard pattern.
+
+    .PARAMETER Name
+        Specifies the name or wildcard pattern of the branch to search for.
+        If not provided, all branch names will be returned.
+
+    .PARAMETER Current
+        Indicates whether to retrieve the name of the current branch. If this switch parameter is present, the function will return the name of the current branch.
 
     .OUTPUTS
-        [System.String]
+        System.String
 
-        The function returns the name of the current local Git branch as a string.
+        The name of the local Git branch.
 
     .EXAMPLE
-        PS C:\> Get-GitLocalBranchName -Current
+        PS> Get-GitLocalBranchName -Name 'main'
 
-        Returns the name of the current local Git branch.
+        Returns the branch that match exactly to the name 'main'.
+
+    .EXAMPLE
+        PS> Get-GitLocalBranchName -Name 'f/*'
+
+        Returns the names of all branches that match the wildcard pattern "f/*".
+
+    .EXAMPLE
+        PS> Get-GitLocalBranchName -Current
+
+        Returns the name of the current Git branch.
+
+    .NOTES
+        This function requires Git to be installed and accessible from the command line.
 #>
 function Get-GitLocalBranchName
 {
-    [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSShouldProcess', '', Justification = 'ShouldProcess is implemented without ShouldProcess/ShouldContinue.')]
-    [CmdletBinding(SupportsShouldProcess = $true, ConfirmImpact = 'Low')]
+    [CmdletBinding()]
     [OutputType([System.String])]
     param
     (
+        [Parameter()]
+        [System.String]
+        $Name,
+
         [Parameter()]
         [System.Management.Automation.SwitchParameter]
         $Current
@@ -31,39 +54,33 @@ function Get-GitLocalBranchName
 
     $branchName = $null
 
-    if ($WhatIfPreference)
+    if ($Current.IsPresent)
     {
-        if ($Current.IsPresent)
-        {
-            Write-Information -MessageData 'What if: Getting current local branch name' -InformationAction Continue
-        }
-        else
-        {
-            Write-Information -MessageData 'What if: Getting local branch names.' -InformationAction Continue
-        }
+        $branchName = git rev-parse --abbrev-ref HEAD
     }
     else
     {
-        if ($Current.IsPresent)
+        if ($Name)
         {
-            $branchName = git rev-parse --abbrev-ref HEAD
+            # Can do wildcard search for branch name, e.g. f/* to find feature branches.
+            $branchName = git branch --format='%(refname:short)' --list $Name
         }
         else
         {
             $branchName = git branch --format='%(refname:short)' --list
         }
+    }
 
-        if ($LASTEXITCODE -ne 0) # cSpell: ignore LASTEXITCODE
-        {
-            $PSCmdlet.ThrowTerminatingError(
-                [System.Management.Automation.ErrorRecord]::new(
-                    $script:localizedData.Get_GitLocalBranchName_Failed,
-                    'GGLBN0001', # cspell: disable-line
-                    [System.Management.Automation.ErrorCategory]::ObjectNotFound,
-                    $branchName
-                )
-            )
+    if ($LASTEXITCODE -ne 0) # cSpell: ignore LASTEXITCODE
+    {
+        $errorMessageParameters = @{
+            Message = $script:localizedData.Get_GitLocalBranchName_Failed
+            Category = 'ObjectNotFound'
+            ErrorId = 'GGLBN0001' # cspell: disable-line
+            TargetObject = $null
         }
+
+        Write-Error @errorMessageParameters
     }
 
     return $branchName