+++ source/Public/Test-GitRemoteBranch.ps1
@@ -6,14 +6,14 @@
         The Test-GitRemoteBranch command checks if a specified branch exists in a
         remote Git repository.
 
-    .PARAMETER BranchName
+    .PARAMETER Name
         Specifies the name of the branch to check.
 
     .PARAMETER RemoteName
         Specifies the name of the remote repository.
 
     .EXAMPLE
-        Test-GitRemoteBranch -BranchName "feature/branch" -RemoteName "origin"
+        Test-GitRemoteBranch -BranchName "feature/branch" -Name "origin"
 
         This example tests if the branch "feature/branch" exists in the remote repository
         named "origin".
@@ -29,28 +29,39 @@
 #>
 function Test-GitRemoteBranch
 {
-    [CmdletBinding()]
+    [CmdletBinding(DefaultParameterSetName = 'Default')]
     [OutputType([System.Boolean])]
     param
     (
-        [Parameter(Mandatory = $true, Position = 0)]
+        [Parameter(Position = 0, ParameterSetName = 'Default')]
+        [Parameter(Mandatory = $true, Position = 0, ParameterSetName = 'Name')]
+        [ValidateNotNullOrEmpty()]
         [System.String]
-        $BranchName,
+        $RemoteName,
 
-        [Parameter(Mandatory = $true, Position = 1)]
+        [Parameter(Position = 1, ParameterSetName = 'Name')]
+        [ValidateNotNullOrEmpty()]
         [System.String]
-        $RemoteName
+        $Name
     )
 
-    # Get the remote branches
-    $branch = ls-remote --branches $RemoteName $BranchName
+    $getGitRemoteBranchParameters = @{}
 
-    $oid, $heads = $branch -split "`t"
+    if ($PSBoundParameters.ContainsKey('RemoteName'))
+    {
+        $getGitRemoteBranchParameters['RemoteName'] = $RemoteName
+    }
+
+    if ($PSBoundParameters.ContainsKey('Name'))
+    {
+        $getGitRemoteBranchParameters['Name'] = $Name
+    }
+
+    $branch = Get-GitRemoteBranch @getGitRemoteBranchParameters -RemoveRefsHeads
 
     $result = $false
 
-    # Check if the branch exists
-    if ($heads -match "^refs/heads/$BranchName")
+    if ($branch)
     {
         $result = $true
     }